Consolidate tests.
